Prepare for refactoring - SimpleAction class created and used for all existing actions
