+++ src/bots/actions/BaseAction.ts
@@ -3,40 +3,15 @@ import * as R from "../../constants/messages";
 import { ProcessMessageSession } from "../events/ProcessMessage";
 
 export default abstract class BaseAction {
-  protected abstract regexp: RegExp | null;
-  private args: string[] | null;
-
-  public testAndExecute(session: ProcessMessageSession) {
-    const message = session.context.message;
-
-    if (this.regexp == null) {
-      throw new Error("Regexp cannot be null");
-    }
-
-    const regexpResults = this.regexp.exec(message.text);
-
-    if (regexpResults == null) {
+  public async testAndExecute(session: ProcessMessageSession): Promise<boolean> {
+    if (!this.test(session))
       return false;
-    }
 
-    regexpResults.shift();
-    return this.execute(session, regexpResults);
+    return this.execute(session);
   }
 
-  public execute(session: ProcessMessageSession, args: string[] | null) {
-    this.args = args;
-    return this.action(session);
-  }
-
-  protected abstract action(session: ProcessMessageSession): Promise<boolean>;
-
-  protected arg(index: number) {
-    if (this.args == null) {
-      throw new Error("Args is null");
-    }
-
-    return this.args[index];
-  }
+  public abstract test(session: ProcessMessageSession): boolean;
+  public abstract execute(session: ProcessMessageSession): Promise<boolean>;
 
   protected processingMessageDelay(session: ProcessMessageSession) {
     const delayInstance = delay(600);
